refactor: add validation and refactor double null-checking on service

+++ src/main/java/site/udtk/sentenceapi/repository/CategoryRepository.java
@@ -12,10 +12,10 @@ public interface CategoryRepository extends JpaRepository<Category, Long> {
 	@Query(value = "SELECT s.id FROM sentence s "
 		+ "JOIN category c ON s.category_id = c.id "
 		+ "WHERE c.language = :language ORDER BY RAND() LIMIT :count", nativeQuery = true)
-	List<Long> findRandomSentenceIdsByLanguage(@Param("language") String language, @Param("count") int count);
+	List<Long> findRandomSentenceIdsByLanguage(@Param("language") String language, @Param("count") Long count);
 
 	@Query(value = "SELECT s.id FROM sentence s "
 		+ "JOIN category c ON s.category_id = c.id "
 		+ "WHERE c.sort = :sort ORDER BY RAND() LIMIT :count", nativeQuery = true)
-	List<Long> findRandomSentenceIdsBySort(@Param("sort") String sort, @Param("count") int count);
+	List<Long> findRandomSentenceIdsBySort(@Param("sort") String sort, @Param("count") Long count);
 }
